+++ game/src/dlls/objects/334_CFPrisonGua/CFPrisonGua.c
@@ -53,7 +53,7 @@ int cfPrisonGuard_sequenceCallback(GameObject* obj, int unused, ObjSeqState* ani
     s8 guardianFreed;
     s8 shouldTransition;
     f32 distance;
-    u32 messageSender;
+    uintptr_t messageSender;
     u32 message;
     u32 messageArgument = 0;
     CfPrisonGuardPlacement* placement = (CfPrisonGuardPlacement*)obj->anim.placement;
